pyproject.toml
[project]
name = "adkbot"
version = "0.1.0"
description = "A multi-model AI assistant framework powered by Google ADK"
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "adkbot contributors" }]
keywords = ["ai", "agent", "chatbot", "adk", "google-adk", "litellm", "multi-model"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    # Core — Google ADK + LiteLLM
    "google-adk>=1.28.1",
    "litellm>=1.83.0",
    "python-dotenv>=1.0.1",
    # Framework
    "pydantic>=2.12.0",
    "httpx>=0.28.0",
    "loguru>=0.7.3",
    "rich>=14.0.0",
    # CLI
    "typer>=0.23.0",
    "prompt-toolkit>=3.0.52",
    "questionary>=2.1.1",
    # Tools
    "ddgs>=9.12.1",
    "readability-lxml>=0.8.4.1",
    "croniter>=6.2.2",
    "mcp>=1.27.0",
    "json-repair>=0.58.7",
    "chardet>=7.4.0.post2",
    "tiktoken>=0.12.0",
    # Channels
    "websocket-client>=1.9.0",
    "dingtalk-stream>=0.24.3",
    "python-telegram-bot[socks]>=22.7",
    "lark-oapi>=1.5.3",
    "socksio>=1.0.0",
    "python-socketio>=5.16.1",
    "msgpack>=1.1.2",
    "slack-sdk>=3.41.0",
    "slackify-markdown>=0.2.2",
    "qq-botpy>=1.2.1",
    "python-socks[asyncio]>=2.8.1",
]

[project.optional-dependencies]
api = ["aiohttp>=3.9.0"]
discord = ["discord.py>=2.5.2"]
matrix = [
    "matrix-nio>=0.25.2",
    "mistune>=3.0.0",
    "nh3>=0.2.17",
]
wecom = ["wecom-aibot-sdk-python>=0.1.5"]
weixin = [
    "qrcode[pil]>=8.0",
    "pycryptodome>=3.20.0",
]

[project.scripts]
adkbot = "adkbot.cli.commands:app"

# --- Build ---
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["adkbot"]

[tool.hatch.build.targets.wheel.force-include]
"bridge" = "adkbot/bridge"

[tool.hatch.build.targets.sdist]
include = ["adkbot/", "bridge/", "README.md", "LICENSE"]

# --- Linting & Testing ---
[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["adkbot"]
omit = ["tests/*", "**/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.",
    "if TYPE_CHECKING:",
]

# --- Modern Dev Dependencies (PEP 735) ---
[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=6.0.0",
    "aiohttp>=3.9.0",
    "ruff>=0.1.0",
]